requirements.txt
Django
djangorestframework
django-cors-headers
pytz
model-mommy